Fixed bugs in memory_cache, some pep8, renamed test to fix naming conflict, added common test class
